fix warning C4244: conversion from 'const double' to 'float', possible loss of data

+++ src/Base/Matrix.h
@@ -309,13 +309,15 @@ inline void Matrix4D::multVec(const Vector3d & src, Vector3d & dst) const
 
 inline void Matrix4D::multVec(const Vector3f & src, Vector3f & dst) const
 {
-  float x = (dMtrx4D[0][0]*src.x + dMtrx4D[0][1]*src.y +
-             dMtrx4D[0][2]*src.z + dMtrx4D[0][3]);
-  float y = (dMtrx4D[1][0]*src.x + dMtrx4D[1][1]*src.y +
-             dMtrx4D[1][2]*src.z + dMtrx4D[1][3]);
-  float z = (dMtrx4D[2][0]*src.x + dMtrx4D[2][1]*src.y +
-             dMtrx4D[2][2]*src.z + dMtrx4D[2][3]);
-  dst.Set(x,y,z);
+  double x = (dMtrx4D[0][0]*src.x + dMtrx4D[0][1]*src.y +
+              dMtrx4D[0][2]*src.z + dMtrx4D[0][3]);
+  double y = (dMtrx4D[1][0]*src.x + dMtrx4D[1][1]*src.y +
+              dMtrx4D[1][2]*src.z + dMtrx4D[1][3]);
+  double z = (dMtrx4D[2][0]*src.x + dMtrx4D[2][1]*src.y +
+              dMtrx4D[2][2]*src.z + dMtrx4D[2][3]);
+  dst.Set(static_cast<float>(x),
+          static_cast<float>(y),
+          static_cast<float>(z));
 }
 
 inline bool Matrix4D::operator== (const Matrix4D& rclMtrx) const
